+++ src/network_manager/network_manager.h
@@ -59,7 +59,7 @@ class NetworkManager{
 		void ReqReceiveThread();
 		void MainThread();
 		void AckThread();
-		void SubscribeNetworkThread(int , int, char*, int);
+		void SubscribeNetworkThread(int , int, char*, uint16_t);
 
 		folly::MPMCQueue<std::optional<struct NetworkRequest>> requestQueue_;
 		folly::MPMCQueue<std::optional<struct NetworkRequest>> ackQueue_;
@@ -71,15 +71,17 @@ class NetworkManager{
 		std::atomic<int> thread_count_{0};
 		bool stop_threads_ = false;
 
-		absl::flat_hash_map<size_t, int> ack_connections_; // <client_id, ack_sock>
+		absl::flat_hash_map<uint16_t, int> ack_connections_; // <client_id, ack_sock>
 		absl::Mutex ack_mu_;
 		absl::Mutex sub_mu_;
-		absl::flat_hash_map<int /* client_id */, std::unique_ptr<SubscriberState>> sub_state_;
+		absl::flat_hash_map<uint16_t /* client_id */, std::unique_ptr<SubscriberState>> sub_state_;
 		int ack_efd_;
 		int ack_fd_ = -1;
 
 		CXLManager *cxl_manager_;
 		DiskManager *disk_manager_;
+
+	
 };
 
 } // End of namespace Embarcadero
